added detailed comments + much convenient testing

+++ src/custom_file.cpp
@@ -1,16 +1,17 @@
 #include <custom_file.h>
 #include <filesystem>
+#include <iostream>
 
 customFile::customFile(
           const std::string& filename,
           std::ios_base::openmode openmode
 )
-          : filename_(filename)
-          , file_(filename, openmode)
+          : filename_{ filename }
+          , file_{ filename, openmode }
 {
-     if (!file_)
+     if (!file_) // если файл не открылся - выбрасываем исключение:
      {
-          std::runtime_error {"File:\t" + filename + " is invalid"};
+          throw std::runtime_error {"File:\t" + filename + " is invalid"};
      }
 }
 
@@ -20,27 +21,27 @@ void customFile::writeToFile(
      responseToObservers response
 )
 {
-     file_.clear();
+     file_.clear(); // очищаем флаги
 
-     if (!file_.is_open())
+     if (!file_.is_open()) // если файл не открыт, то
      {
-          file_.open(filename_, openmode);
+          file_.open(filename_, openmode); // открываем файл для записи с соответсвующим флагом
      }
 
-     file_ << text;
+     file_ << text; // записываем данные в файл
 
      if (response == responseToObservers::notification_on)
      {
-          update();
+          update(); // оповещаем наблюдателей
      }
 }
 
 void customFile::closeFile(responseToObservers response)
 {
-     file_.close();
+     file_.close(); // закрываем файл
      if (response == responseToObservers::notification_on)
      {
-          update();
+          update(); // оповещаем наблюдателей
      }
 }
 
@@ -50,29 +51,35 @@ void customFile::openFile(
      responseToObservers response
 )
 {
-     if (filename != filename_)
+     if (filename != filename_) // если имя текущего файла != имени,
+                              // указанного в аргументе, то
      {
-          file_.close();
-          filename_ = filename;
+          file_.close(); // закрываем текущий файл
+          filename_ = filename; // обновляем имя текущего файла
      }
 
-     file_.open(filename, openmode);
+     file_.open(filename, openmode); // открываем файл с нужным флагом открытия
      if (response == responseToObservers::notification_on)
      {
-          update();
+          update(); // оповещаем наблюдателей
      }
 }
 
 bool customFile::removeFile(responseToObservers response)
 {
-     bool is_removed = std::filesystem::remove(filename_);
-
-     if (is_removed)
+     if (!std::filesystem::exists(filename_)) // если не существует
      {
-          filename_ = "";
+          std::cout << "Cant remove: file doest not exists" << std::endl;
+          return false;
+     }
+
+     bool is_removed = std::filesystem::remove(filename_); // удаляем файл
+
+     if (is_removed) // если файл удалился, то
+     {
           if (response == responseToObservers::notification_on)
           {
-               update();
+               update(); // оповещаем наблюдателей
           }
      }
      return is_removed;
@@ -90,7 +97,7 @@ customFile::~customFile()
 
 void customFile::update()
 {
-     notify(*this);
+     notify(*this); // вызываем метод, определенный в базовом классе Subject для оповещения наблюдателей
 }
 
 customFile::customFile()
